+++ src/main/java/com/codeborne/selenide/conditions/OwnText.java
@@ -1,8 +1,6 @@
 package com.codeborne.selenide.conditions;
 
 import com.codeborne.selenide.Driver;
-import com.codeborne.selenide.TextCheck;
-import com.codeborne.selenide.impl.Html;
 import org.openqa.selenium.WebElement;
 
 import javax.annotation.CheckReturnValue;
@@ -13,27 +11,13 @@ import static com.codeborne.selenide.commands.GetOwnText.getOwnText;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 
 @ParametersAreNonnullByDefault
-public class OwnText extends TextCondition {
+public class OwnText extends CaseInsensitiveTextCondition {
   public OwnText(String expectedText) {
     super("own text", expectedText);
     if (isEmpty(expectedText)) {
       throw new IllegalArgumentException("Argument must not be null or empty string. " +
         "Use $.shouldHave(exactOwnText(\"\").");
     }
-  }
-
-  @CheckReturnValue
-  @Override
-  protected boolean match(String actualText, String expectedText) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  protected boolean match(TextCheck textCheck, String actualText, String expectedText) {
-    return switch (textCheck) {
-      case FULL_TEXT -> Html.text.equals(actualText, expectedText);
-      case PARTIAL_TEXT -> Html.text.contains(actualText, expectedText);
-    };
   }
 
   @Nullable
