+++ src/main/java/com/codeborne/selenide/conditions/Text.java
@@ -1,8 +1,6 @@
 package com.codeborne.selenide.conditions;
 
 import com.codeborne.selenide.Driver;
-import com.codeborne.selenide.TextCheck;
-import com.codeborne.selenide.impl.Html;
 import org.openqa.selenium.WebElement;
 
 import javax.annotation.CheckReturnValue;
@@ -12,7 +10,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 
 @ParametersAreNonnullByDefault
-public class Text extends TextCondition {
+public class Text extends CaseInsensitiveTextCondition {
 
   public Text(final String text) {
     super("text", text);
@@ -21,20 +19,6 @@ public class Text extends TextCondition {
       throw new IllegalArgumentException("Argument must not be null or empty string. " +
         "Use $.shouldBe(empty) or $.shouldHave(exactText(\"\").");
     }
-  }
-
-  @CheckReturnValue
-  @Override
-  protected boolean match(String actualText, String expectedText) {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
-  protected boolean match(TextCheck textCheck, String actualText, String expectedText) {
-    return switch (textCheck) {
-      case FULL_TEXT -> Html.text.equals(actualText, expectedText);
-      case PARTIAL_TEXT -> Html.text.contains(actualText, expectedText);
-    };
   }
 
   @Nullable
